# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PC Components Warehouse Management API
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export interface Component {
  id: number;
  category: string;
  model: string;
  quantity: number;
  price: number;
  cost: number;
  imageUrl?: string | null;
  /** Profit margin as a percentage (0-100) */
  marginPct: number;
  createdAt: string;
}

export interface CreateComponent {
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  model: string;
  /** @minimum 0 */
  quantity: number;
  /** @minimum 0.01 */
  price: number;
  /** @minimum 0 */
  cost?: number;
  imageUrl?: string;
}

export interface UpdateComponent {
  /** @minimum 0.01 */
  price?: number;
  /** @minimum 0 */
  cost?: number;
  imageUrl?: string | null;
  /** Quantity to add (can be negative to deduct) */
  addQuantity?: number;
  /**
   * Set absolute stock level (overrides addQuantity)
   * @minimum 0
   */
  setQuantity?: number;
}

export interface BulkIds {
  /** @minItems 1 */
  ids: number[];
}

export interface BulkOperationResult {
  deleted: number;
}

export interface Sale {
  id: number;
  componentId: number;
  category: string;
  model: string;
  quantity: number;
  unitPrice: number;
  unitCost: number;
  total: number;
  totalCost: number;
  profit: number;
  customerName?: string | null;
  isReturned: boolean;
  returnedAt?: string | null;
  soldAt: string;
}

export interface CreateSale {
  componentId: number;
  /** @minimum 1 */
  quantity: number;
  customerName?: string;
}

export interface Suggestion {
  model: string;
  approxPrice: number;
  approxCost: number;
}

export interface DashboardSummary {
  totalItems: number;
  totalUnits: number;
  totalValue: number;
  totalCostValue: number;
  lowStockCount: number;
  lowStockThreshold: number;
  todayRevenue: number;
  todayProfit: number;
  todayTransactions: number;
  totalRevenue: number;
  totalProfit: number;
  marginPct: number;
  totalTransactions: number;
  topCategory?: string | null;
}

export interface DailyRevenuePoint {
  date: string;
  revenue: number;
  profit: number;
  transactions: number;
}

export interface CategoryRevenuePoint {
  category: string;
  revenue: number;
  profit: number;
  units: number;
}

export interface ProductPerformance {
  componentId: number;
  category: string;
  model: string;
  unitsSold: number;
  revenue: number;
  profit: number;
  currentStock: number;
}

export interface ReorderItem {
  id: number;
  category: string;
  model: string;
  currentStock: number;
  sold30d: number;
  daysOfStock?: number | null;
  /** critical | low | ok */
  urgency: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListComponentsParams = {
  category?: string;
  search?: string;
};

export type GetLowStockParams = {
  threshold?: number;
};

export type GetBestSellersParams = {
  limit?: number;
};

export type GetSlowMoversParams = {
  limit?: number;
};