# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PC Components Warehouse Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /components:
    get:
      operationId: listComponents
      tags: [components]
      summary: List inventory components
      parameters:
        - in: query
          name: category
          required: false
          schema:
            type: string
        - in: query
          name: search
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Components list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Component"
    post:
      operationId: createComponent
      tags: [components]
      summary: Add new component (auto-assigns ID)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComponent"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Component"
  /components/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    patch:
      operationId: updateComponent
      tags: [components]
      summary: Update price and/or stock for a component
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComponent"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Component"
    delete:
      operationId: deleteComponent
      tags: [components]
      summary: Remove a component
      responses:
        "204":
          description: Deleted
  /components/bulk-delete:
    post:
      operationId: bulkDeleteComponents
      tags: [components]
      summary: Delete multiple components by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkIds"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkOperationResult"
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: Sales history (newest first)
      responses:
        "200":
          description: Sales list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: recordSale
      tags: [sales]
      summary: Record a sale (decrements stock)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSale"
      responses:
        "201":
          description: Sale recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}/return:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    post:
      operationId: returnSale
      tags: [sales]
      summary: Mark a sale returned and restock the component
      responses:
        "200":
          description: Returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
...
  /analytics/low-stock:
    get:
      operationId: getLowStock
      tags: [analytics]
      summary: Components with stock below threshold
      parameters:
        - in: query
          name: threshold
          required: false
          schema:
            type: integer
            default: 3
      responses:
        "200":
          description: Low stock items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Component"
  /analytics/best-sellers:
    get:
      operationId: getBestSellers
      tags: [analytics]
      summary: Best-selling components by units sold
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Best sellers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductPerformance"
  /analytics/slow-movers:
    get:
      operationId: getSlowMovers
      tags: [analytics]
      summary: Items with low/no recent sales
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Slow movers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductPerformance"
...
components:
...
  schemas:
...
    Component:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        model:
          type: string
        quantity:
          type: integer
        price:
          type: number
        cost:
          type: number
        imageUrl:
          type: string
          nullable: true
        marginPct:
          type: number
          description: Profit margin as a percentage (0-100)
        createdAt:
          type: string
          format: date-time
      required: [id, category, model, quantity, price, cost, marginPct, createdAt]
    CreateComponent:
      type: object
      properties:
        category:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        quantity:
          type: integer
          minimum: 0
        price:
          type: number
          minimum: 0.01
        cost:
          type: number
          minimum: 0
        imageUrl:
          type: string
      required: [category, model, quantity, price]
    UpdateComponent:
      type: object
      properties:
        price:
          type: number
          minimum: 0.01
        cost:
          type: number
          minimum: 0
        imageUrl:
          type: string
          nullable: true
        addQuantity:
          type: integer
          description: Quantity to add (can be negative to deduct)
        setQuantity:
          type: integer
          minimum: 0
          description: Set absolute stock level (overrides addQuantity)
    BulkIds:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
          minItems: 1
      required: [ids]
    BulkOperationResult:
      type: object
      properties:
        deleted:
          type: integer
      required: [deleted]
    Sale:
      type: object
      properties:
        id:
          type: integer
        componentId:
          type: integer
        category:
          type: string
        model:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        unitCost:
          type: number
        total:
          type: number
        totalCost:
          type: number
        profit:
          type: number
        customerName:
          type: string
          nullable: true
        isReturned:
          type: boolean
        returnedAt:
          type: string
          format: date-time
          nullable: true
        soldAt:
          type: string
          format: date-time
      required: [id, componentId, category, model, quantity, unitPrice, unitCost, total, totalCost, profit, isReturned, soldAt]
    CreateSale:
      type: object
      properties:
        componentId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        customerName:
          type: string
      required: [componentId, quantity]
...
    ProductPerformance:
      type: object
      properties:
        componentId:
          type: integer
        category:
          type: string
        model:
          type: string
        unitsSold:
          type: integer
        revenue:
          type: number
        profit:
          type: number
        currentStock:
          type: integer
      required: [componentId, category, model, unitsSold, revenue, profit, currentStock]
...