# lib/api-zod/src/generated/types/bulkOperationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PC Components Warehouse Management API
 * OpenAPI spec version: 0.1.0
 */

export interface BulkOperationResult {
  deleted: number;
}