# lib/api-zod/src/generated/types/createComponent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PC Components Warehouse Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateComponent {
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  model: string;
  /** @minimum 0 */
  quantity: number;
  /** @minimum 0.01 */
  price: number;
  /** @minimum 0 */
  cost?: number;
  imageUrl?: string;
}